runfix: Address revoked devices not being checked
