+++ src/script/E2EIdentity/mlsStatus.ts
@@ -17,21 +17,22 @@
  *
  */
 
-import {WireIdentity} from '@wireapp/core/lib/messagingProtocols/mls';
+import {DeviceStatus} from '@wireapp/core/lib/messagingProtocols/mls';
 
-import {MLSStatuses} from './E2EIdentityVerification';
+export enum MLSStatuses {
+  VALID = 'valid',
+  NOT_ACTIVATED = 'not_activated',
+  EXPIRED = 'expired',
+  EXPIRES_SOON = 'expires_soon',
+  REVOKED = 'revoked',
+}
 
-type CoreStatus = WireIdentity['status'];
+const statusMap: Record<DeviceStatus, MLSStatuses> = {
+  [DeviceStatus.Valid]: MLSStatuses.VALID,
+  [DeviceStatus.Expired]: MLSStatuses.EXPIRED,
+  [DeviceStatus.Revoked]: MLSStatuses.REVOKED,
+};
 
-export const mapMLSStatus = (status?: CoreStatus) => {
-  const statusMap: Record<any, MLSStatuses> = {
-    Valid: MLSStatuses.VALID,
-    Expired: MLSStatuses.EXPIRED,
-    Revoked: MLSStatuses.REVOKED,
-  };
-
-  if (!status) {
-    return MLSStatuses.NOT_ACTIVATED;
-  }
-  return statusMap[status];
+export const mapMLSStatus = (status?: DeviceStatus) => {
+  return !status ? MLSStatuses.NOT_ACTIVATED : statusMap[status];
 };
